backend/requirements.txt


# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Semantic Kernel
semantic-kernel

# AI/ML
google-generativeai==0.3.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Database
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
tenacity==8.2.3

# Logging
loguru==0.7.2

numpy>=1.24.0